requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
pymongo==4.6.0
redis==5.0.1
motor==3.3.2

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
cryptography==41.0.7

# ML/AI Libraries
tensorflow==2.15.0
torch==2.1.1
transformers==4.36.0
sentence-transformers==2.2.2
scikit-learn==1.3.2
numpy==1.24.3
pandas==2.1.4
scipy==1.11.4

# Vector Database
pinecone-client==2.2.4
weaviate-client==3.25.3
faiss-cpu==1.7.4

# Search & NLP
elasticsearch==8.11.0
spacy==3.7.2
nltk==3.8.1
gensim==4.3.2

# Schema & Knowledge Graph
rdflib==7.0.0
networkx==3.2.1
py2neo==2021.2.3

# Message Queue
pika==1.3.2
celery==5.3.4
kombu==5.3.4

# Monitoring & Analytics
prometheus-client==0.19.0
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0

# API Clients
httpx==0.25.2
aiohttp==3.9.1
requests==2.31.0

# Data Processing
beautifulsoup4==4.12.2
lxml==4.9.3
html2text==2020.1.16
markdown==3.5.1
pypdf2==3.0.1
pillow==10.1.0

# Web Crawling & Content Processing
playwright==1.40.0
selenium==4.16.0
python-magic==0.4.27
chardet==5.2.0
langdetect==1.0.9
ftfy==6.1.3
extruct==0.16.0
w3lib==2.1.2
datasketch==1.6.4
simhash==2.1.2
python-docx==1.1.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
faker==20.1.0

# Development Tools
black==23.11.0
pylint==3.0.2
mypy==1.7.1
pre-commit==3.5.0
pyotp==2.9.0

# API Documentation
mkdocs==1.5.3
mkdocs-material==9.4.14

# Scheduling
apscheduler==3.10.4

# Cloud SDKs
boto3==1.34.0
google-cloud-storage==2.10.0
azure-storage-blob==12.19.0

# Rate Limiting
slowapi==0.1.9

# Caching
aiocache==0.12.2

# Validation
jsonschema==4.20.0
pydantic-settings==2.1.0

# Logging
structlog==23.2.0
python-json-logger==2.0.7

# Performance
ujson==5.8.0
orjson==3.9.10

# GraphQL
strawberry-graphql[fastapi]==0.215.1

# WebSockets
websockets==12.0
python-socketio==5.10.0

# Feature Flags
unleash-client-python==5.9.0

# Distributed Tracing
jaeger-client==4.8.0

# Configuration Management
dynaconf==3.2.4